Introduce console_task.
This is a new task pointer, that under all normal circumstances is
equivalent to current_task. However, on ISR/IRQ handler entry,
it's changed to &kernel_task, making putchar() et al. output
to the kernel console, no matter which task was interrupted to
get the kernel code running. Previously, "force_current_console"
did something similar, but that turned out to be a very ugly hack.