Fix a truly ancient bug: buffer overflow in the physical memory allocator.
This thing has haunted me for a few days now... When creating tasks
for user mode, I noticed that a list became corrupt. Testing further,
it became clear that something was writing 0xff bytes to memory.
It turned out to be the physical memory allocator writing used - 0xff
simply means 8 frames in row that are used...