Make the userspace address space continuous; remove unneeded parameter from vmm calls
Userspace now uses 0x10000000 for loading ELF stuff (as before), and 0xbffff000 and downwards
as the userspace stack. This makes the userspace address range just that - a range,
from 0x10000000 to 0xbfffffff - with everything else (before and after) being kernel space.

Also remove the "kernelspace" parameter from _vmm_create_page_table and _vmm_get_page,
as that functionality can now be taken care of with a simple macro.