Edit Makefile to supporting simics (and add simics.sh for Linux)
Unfortunately, this means using GCC instead of Clang, as I can't
get Clang to emit STABS debug symbols. Ugh! I suppose I could
use Clang for everyday stuff and GCC for simics compilation,
but that would mean using a different compiler when trying to
find hard-to-find bugs - who knows if the output is similar enough
to even trigger the same bugs? It should be, of course...
Ramble ramble.