Change list_remove into sensible functions.
It was the only function that took a node as an argument, rather than a data
member the user had supplied. Therefore, all remove() calls minus one were
of the form list_remove(list, list_find_first(list, data)).
There is now list_remove_node, which is the above function renamed, and
list_remove_{first,last} that take a list and a data member as an argument.