Add a user-mode shell!
It's a port of a basic shell I wrote recently, with the intention of eventually
porting it. The main cause of writing it then, however, was to get better at
UNIX programming, including fork, exec*, dup/dup2 for redirection etc.

It's still basic - no redirection or pipes, etc. Next up: solve execve memory leaks
on failure.