I FINALLY managed to read a full, large ext2 directory. MESSY CODE.
Pre-cleanup commit, so that I don't manage to screw something up
without the ability to revert..

In short, the problem was that there are unused directory entries in the
middle of a directory, that you are supposed to *skip* (ahead by 1 blocksize),
but I never saw anything about that in the docs; I had to figure it own myself.
My initial assumption was that you should *stop reading* when you find such an entry.
Instead, you need to calculate (from i_blocks) how many blocks the directory
consists of, and **keep reading no matter what** until you've checked them all.

The part that made me perplexed is that the blocklist in the inode actually points
to these **unused** directory entry blocks as if they are part of the directory, yet
they contain nonsense information (apart from the first four bytes, the inode number,
being zero). I suspected disk/image corruption before I realized what the heck was
going on!