Fix assertion failure in physical memory manager.
Long story short (part of it is in the code, as a comment):
Some "computers (VMware Fusion, in this case) has non-continous memory, near
the end of RAM. exscapeOS uses a bitmap to keep track of allocated frames,
and the size of that bitmap was calculated using the size of the *continous*
memory past 1 MB. If there was additional memory after a discontinuity,
the bitmap's size was incorrectly calculated.

This was fixed by a fairly ugly loop that checks all the memory areas found
by GRUB, and keeps track of the highest address. The bitmap is then sized
such that in can keep track of all frames between 0x0 and that maximum address.