Fix ATA driver bug that caused hangs on real computers.
After sending the IDENTIFY command, there was no wait before
reading the status register, which caused nonexistent drives
to be found by the OS (as they didn't have time to respond).
Later on, when waiting for DRQ to be set, the OS hanged forever,
as a nonexistent device will never become ready for data transfer!