Add lstat, readlink; TONS of ext2 work (see full msg)
This was not really intended to be added as one commit, but the amount of work
has become so great that I simply cannot finish everything without committing
the current state as a backup. The work has progressed for many work hours with
no version control, which is not good at all.

ext2_read() (i.e. read() on ext2 filesystems) does not work yet.
Most of the recent work has been on getting symlinks to work.
As of right now, they mostly work. A few known flaws:
* Symlinks aren't expanded in paths, so e.g. with a symlink named "etc" to /etc,
  the path "etc/.." should point to the root directory, but it will instead point
  to the directory where the symlink is located, as the path is not expanded
  prior to the .. "operator" acting.
* Symlink loops are not detected, and an infinite link loop will probably either
  hang or panic (I'm not sure off hand whether it'd cause a stack overflow or just
  an infinite loop in one stack frame).
* A few bonus features like a O_SYMLINK flag to open() are not supported yet.