Expand the userspace stack as needed!
Previously, the userspace stack was limited to 16 kiB.
Now, it's 16 kB by default, and increases as needed, by 16 kiB,
until the maximum of 8 MiB.

There is a big potential issue remaining, though: sometimes, resizing
requires a new page table to be allocated, which requires kmalloc...
which does not work when called from an ISR (such as page_fault_handler).
I'm about to look at this now.