Allow kmalloc from ISRs (see msg)
This solves the issue where a userspace stack can't grow beyond ~4 MB,
as a new page table needs to be allocated at that point.

As far as I can tell/remember, there is no reason kmalloc shouldn't work in ISRs now.
I believe the reason I disallowed it earlier was because it used mutexes, which
only work if the scheduler is active. For various reasons, it now uses interrupt locking.