Hack-add . and .. for initrd; change initrd root to be inode 1.
While adding in "." and "..", I realized that using inode 0 for the root
wasn't a great idea; Newlib's readdir actively ignores listing entries with
inode number 0. I changed the root to be inode 1 to take care of this; also,
inode 0 in the initrd is now a null entry (to keep the indexes and memory locations
in sync in an easy way).