Basic memory use (RSS) tracing for userspace tasks.
Only visible through ps (which is a kernel shell hack, not a program).
It also doesn't consider stuff allocated using kmalloc, such as the program's
symbols, or the task's kernel stack, and so on.
Memory allocated using malloc in userspace is tracked, via sbrk -> vmm_alloc_user.
Memory copied when fork()ing is also counted for the child task.
(Since copy-on-write fork is not yet supported, all memory is copied on fork.)